requirements.txt
numpy==1.11.0

scikit_image==0.13.1

Pillow==5.1.0

tensorflow==1.8.0
